+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - A fallback value parsing from the \<paint\> type.
+- Parse the `inherit` value from the `font-family` attribute.
 
 ## [0.0.2] - 2016-09-09
 ### Removed